+++ main.go
@@ -5,8 +5,9 @@ import (
 	"log"
 	"os"
 
-	"github.com/xsb/dog/dog"
-	"github.com/xsb/dog/dogfile"
+	"github.com/xsb/dog/executor"
+	"github.com/xsb/dog/parser"
+	"github.com/xsb/dog/types"
 )
 
 func printHelp() {
@@ -14,7 +15,7 @@ func printHelp() {
 	fmt.Println("Dog Help")
 }
 
-func printTasks(tm dog.TaskMap) {
+func printTasks(tm types.TaskMap) {
 	for k, t := range tm {
 		fmt.Printf("%s\t%s\n", k, t.Description)
 	}
@@ -25,7 +26,7 @@ func main() {
 
 	// dog
 	case len(os.Args) == 1:
-		tm, err := dogfile.LoadDogFile()
+		tm, err := parser.LoadDogFile()
 		if err != nil {
 			fmt.Println("Error: No valid Dogfile in current directory")
 			fmt.Println("Need help? --> dog help")
@@ -42,17 +43,17 @@ func main() {
 	case len(os.Args) >= 2 && os.Args[1] != "help":
 		taskName := os.Args[1]
 
-		tm, err := dogfile.LoadDogFile()
+		tm, err := parser.LoadDogFile()
 		if err != nil {
 			log.Fatal(err)
 		}
 
 		if task, ok := tm[taskName]; ok {
-			var e *dog.Executor
+			var e *executor.Executor
 			if task.Executor != "" {
-				e = dog.NewExecutor(task.Executor)
+				e = executor.NewExecutor(task.Executor)
 			} else {
-				e = dog.SystemExecutor
+				e = executor.SystemExecutor
 			}
 
 			if err := e.Exec(&task, os.Stdout); err != nil {
@@ -61,6 +62,5 @@ func main() {
 		} else {
 			fmt.Println("No task named " + taskName)
 		}
-
 	}
 }
