+++ parser/parser.go
@@ -1,4 +1,4 @@
-package dogfile
+package parser
 
 import (
 	"io/ioutil"
@@ -6,18 +6,18 @@ import (
 	"regexp"
 
 	"github.com/ghodss/yaml"
-	"github.com/xsb/dog/dog"
+	"github.com/xsb/dog/types"
 )
 
-func ParseDogfile(d []byte) (tm dog.TaskMap, err error) {
-	var tasks []dog.Task
+func ParseDogfile(d []byte) (tm types.TaskMap, err error) {
+	var tasks []types.Task
 
 	err = yaml.Unmarshal(d, &tasks)
 	if err != nil {
 		return
 	}
 
-	tm = make(dog.TaskMap)
+	tm = make(types.TaskMap)
 	for _, t := range tasks {
 		if _, ok := tm[t.Name]; ok {
 			// TODO (duplicated task name) fail and return a non-nil error
@@ -30,7 +30,7 @@ func ParseDogfile(d []byte) (tm dog.TaskMap, err error) {
 }
 
 // LoadDogFile finds a Dogfile in disk, parses YAML and returns a map
-func LoadDogFile() (tm dog.TaskMap, err error) {
+func LoadDogFile() (tm types.TaskMap, err error) {
 	const validDogfileName = "^(Dogfile|🐕)"
 	var dogfiles []os.FileInfo
 	var d []byte
